requirements.txt
azure-functions
Pillow
azure-storage-blob
